Add native macOS build support (Docker-free)
Running under Docker Desktop for Mac routes LMDB through the VM file-sharing layer
and caps the page cache at the VM RAM, which is slow for queries. This adds a native
macOS build path (mmaps the DB directly, uses full host RAM as cache).

- src/Bytes32.h: cast to uint64_t* (size_t != uint64_t on macOS)
- src/PluginEventSifter.h: st_mtim->st_mtimespec, environ via crt_externs (__APPLE__)
- Makefile: link -luv on macOS (uWebSockets uses libuv, not epoll)
- patches/file_change_monitor.h: kqueue port of the inotify watcher (nested submodule;
  applied by the build script). Preserves instant notification for real-time delivery.
- build-native-macos.sh + docs/native-macos-build.md: one-command build + instructions

All changes are #ifdef/uname-guarded; the Linux/Docker build is unaffected.

+++ src/Bytes32.h
@@ -33,7 +33,7 @@ namespace std {
     template<> struct hash<Bytes32> {
         std::size_t operator()(Bytes32 const &b) const {
             static_assert(sizeof(b.buf) == 32);
-            uint64_t *p = (size_t*)&b.buf;
+            uint64_t *p = (uint64_t*)&b.buf;
             return size_t(p[0] ^ p[1] ^ p[2] ^ p[3]);
         }
     };
